refactor: add ExtendedRequestInit
